+++ changelog.md
@@ -1,4 +1,16 @@
 # Changelog
 
+**./src/common/possible-criteria-message.js**
+* Prefixed with "Invalid route criteria"
+
+---
+
+**./src/common/route-criteria.js - readCriteria**
+* Changed 'reason' for 'type' clause.
+	* Before: "Invalid value type."
+	* After: "Value type not allowed."
+
+---
+
 **./test-parts/test-pathfinding-input.js - handlePossibleRoutes**
-* Wrote "Invalid Criteria Object" unit test stub.
+* Commented out `expect` for "Invalid Criteria Array"